+++ Projet/src/pressuriseur.cpp
@@ -2,6 +2,7 @@
 
 using namespace std;
 
+
 Pressuriseur::Pressuriseur() : m_etat(1.), m_etatResistance(1.), m_tempDemandee(25), m_temp_actuel(25)
 {
 }
@@ -16,12 +17,12 @@ double Pressuriseur::etatResistance() const
   return m_etatResistance;
 }
 
-int Pressuriseur::temperatureDemandee() const
+double Pressuriseur::temperatureDemandee() const
 {
   return m_tempDemandee;
 }
 
-int Pressuriseur::temperatureActuel() const
+double Pressuriseur::temperatureActuel() const
 {
   return m_temp_actuel;
 }
